Rettet tests
Former-commit-id: 879eb8a9c8cb2ede0fe66e9d7cf39d9665508d54

+++ src/rubiks/model.py
@@ -91,7 +91,7 @@ class Model(nn.Module):
 		return new_net
 
 	def get_params(self):
-		return torch.cat([x.flatten() for x in self.state_dict().values()]).clone()
+		return torch.cat([x.float().flatten() for x in self.state_dict().values()]).clone()
 	
 	def save(self, save_dir: str, is_min=False):
 		"""
